git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@372 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/server/forward_messagebuffer.cpp
@@ -72,7 +72,7 @@ bool Forward_MessageBuffer::process()
 	}
 	
 	Components::COMPONENTS cts = Components::getSingleton().getComponents(forwardComponentType_);
-	int32 idx = 0;
+	size_t idx = 0;
 
 	if(cts.size() > 0)
 	{
